+++ src/components/SearchFilter/SearchFilter.module.scss
@@ -16,7 +16,7 @@
         width: 100px;
         cursor: pointer;
         text-align: start;
-        background-image: url('../../../assets/image/toggleDown.png');
+        background-image: url('../../assets/image/toggleDown.png');
         background-repeat: no-repeat;
         background-size: 10px 8px;
         background-position: 100% 50%;
@@ -27,7 +27,7 @@
         width: 100px;
         cursor: pointer;
         text-align: start;
-        background-image: url('../../../assets/image/toggleUp.png');
+        background-image: url('../../assets/image/toggleUp.png');
         background-repeat: no-repeat;
         background-size: 10px 8px;
         background-position: 100% 50%;
