+++ src/pages/SearchPage/SearchPage.scss
@@ -1,3 +1,76 @@
+.searchKeyWordBar {
+  width: 100%;
+  border-bottom: 1px solid #ececec;
+  .searchKeyWord {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    width: 1248px;
+    height: 67px;
+    margin: 0 auto;
+    & .searchLogoImg {
+      width: 18px;
+      height: 18px;
+      background: {
+        image: url('../../assets/image/search.png');
+        size: 18px 18px;
+        repeat: no-repeat;
+      }
+      text-indent: 30px;
+    }
+    & .searchinput {
+      width: 150px;
+      height: 40px;
+      margin-left: 25px;
+      cursor: text;
+      font-size: 13px;
+      border: none;
+    }
+    & .searchinput:focus {
+      outline: none;
+    }
+
+    // .inputResetBtn {
+    //   width: 15px;
+    //   height: 15px;
+    //   display: flex;
+    //   justify-content: center;
+    //   align-items: center;
+    //   background: {
+    //     image: url(../../assets/image/close.png);
+    //     size: 29px 15px;
+    //     position: 0.5px 0;
+    //   }
+    // }
+    // .inputResetBtn:hover {
+    //   cursor: pointer;
+    //   background: {
+    //     image: url(../../assets/image/close.png);
+    //     size: 29px 15px;
+    //     position: -14px 0;
+    //   }
+    // }
+
+    // .blind {
+    //   width: 0;
+    //   height: 0;
+    // }
+  }
+}
+
+.count {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  width: 1248px;
+  height: 67px;
+  margin: 0 auto;
+  font-size: 13px;
+  .workCnt {
+    margin-left: 5px;
+  }
+}
+
 .noResult {
   width: 100%;
   height: 650px;
